dvc_pipelines/sklearn_classifier/requirements.txt
dvc[gdrive]~=2.9.5
numpy~=1.22.2
scikit-learn~=1.0.2
PyYAML~=6.0
pandas~=1.4.1